third_party/WebKit/LayoutTests/fast/css/pseudo-any-expected.txt

This page tests :-webkit-any. This window needs to be focused for the :focus tests to pass.

h1
h1
div
foo
b i

p

a span
li
u s
div span
div~q q+q q>span

first

last



h1
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#div1
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


Element: DIV
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgba(0, 0, 0, 0)'


b
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


i
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


p
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#span1
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgba(0, 0, 0, 0)'


span:focus
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


a
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


a:focus
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 128, 0)'


.foo
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#bar
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


h1.h1class
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 128, 0)'


PASS document.querySelectorAll('#container-div :-webkit-any([href],input[type],input[name])').length is 2
PASS document.querySelectorAll('#container-div :-webkit-any(div,a):-webkit-any([type],[href],[name])').length is 2
PASS document.querySelectorAll('#container-div :-webkit-any(:link, :not(a))').length is 3


#container-div input
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 128, 0)'


#container-div a[href]
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 128, 0)'


#container-div div
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 128, 0)'


#container-div a[name]
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgba(0, 0, 0, 0)'


s
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgba(0, 0, 0, 0)'


div span
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


div ~ q
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


q+q
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


q>span
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#test-last-child :first-child
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#test-last-child :last-child
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
